refactor: Controller, View 작성, 기능 이동
