--- src/main/java/com/mission2/racingcar/controller/RaccingController.java
@@ -1,4 +1,0 @@
-package com.mission2.racingcar.controller;
-
-public class RaccingController {
-}
