+++ src/main/java/com/mission2/racingcar/controller/RacingController.java
@@ -1,17 +1,21 @@
-package com.mission2.racingcar.domain;
+package com.mission2.racingcar.controller;
+
+import com.mission2.racingcar.domain.CarRacingService;
+import com.mission2.racingcar.view.View;
 
 import java.util.List;
-import java.util.Scanner;
 
-public class CarRacing {
+public class RacingController {
+
+    static CarRacingService service = new CarRacingService();
+    static View view = new View();
 
     public static void main(String[] args) {
-        CarRacingService service = new CarRacingService();
 
-        String inputCarNames = getUserInput("경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분");
+        String inputCarNames = view.getUserInput("경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분");
         List<String> carNames = ensureCarNameCondition(service, inputCarNames);
 
-        String inputGameCount = getUserInput("시도할 회수는 몇회인가요?");
+        String inputGameCount = view.getUserInput("시도할 회수는 몇회인가요?");
         int gameCount = ensureGameCountCondition(service, inputGameCount);
 
         service.game(carNames, gameCount);
@@ -21,7 +25,7 @@ public class CarRacing {
         int gameCount = service.getGameCount(inputGameCount);
         while (!service.checkGameCountNotOverMaxGameCount(gameCount)) {
             String message = "게임 횟수는 " + CarRacingService.MAX_GAME_COUNT + "회를 넘을 수 없습니다.";
-            gameCount = service.getGameCount(getUserInput(message));
+            gameCount = service.getGameCount(view.getUserInput(message));
         }
         return gameCount;
     }
@@ -31,16 +35,9 @@ public class CarRacing {
         List<String> carNames = service.splitCarNamesByComma(inputCarNames);
         while (!service.checkCarNamesNotOverMaxCarCount(carNames)) {
             String message = "자동차 이름은 " + CarRacingService.MAX_CAR_NAME_COUNT + "글자를 넘을 수 없습니다.";
-            carNames = service.splitCarNamesByComma(getUserInput(message));
+            carNames = service.splitCarNamesByComma(view.getUserInput(message));
         }
         return carNames;
-    }
-
-    public static String getUserInput(String message) {
-        System.out.println(message);
-        Scanner scanner = new Scanner(System.in);
-
-        return scanner.nextLine();
     }
 
 }
