+++ src/main/java/com/mission2/racingcar/view/View.java
@@ -1,4 +1,13 @@
 package com.mission2.racingcar.view;
 
 public class View {
+
+    InputView inputView = new InputView();
+    OutputView outputView = new OutputView();
+
+    public String getUserInput(String message) {
+        return inputView.getUserInput(message);
+    }
+
+
 }
